refactor: move location services (DD-63)
