+++ src/app/features/admin/services/location-api/location-api.service.ts
@@ -1,16 +1,17 @@
 import { HttpClient } from '@angular/common/http';
 import { inject, Injectable } from '@angular/core';
 import { map } from 'rxjs';
+import { environment } from 'src/environments/environment';
 
 const responseLimit = 5;
 
-export interface NominatimResponse {
+interface NominatimApiResponse {
   display_name: string;
   lat: string;
   lon: string;
 }
 
-interface NominatimResponseCity {
+interface NominatimApiCityResponse {
   address: {
     city?: string;
     town?: string;
@@ -23,11 +24,11 @@ interface NominatimResponseCity {
 export class LocationApiService {
   private readonly httpClient = inject(HttpClient);
 
-  private readonly apiUrl = 'https://nominatim.openstreetmap.org';
+  private readonly apiUrl = environment.baseLocationApiUrl;
 
   public getLocationCoordinates(city: string) {
     const url = `${this.apiUrl}/search?q=${city}&format=json&limit=${responseLimit}`;
-    return this.httpClient.get<NominatimResponse[]>(url).pipe(
+    return this.httpClient.get<NominatimApiResponse[]>(url).pipe(
       map((response) =>
         response.map((cit) => ({
           name: cit.display_name.split(',')[0],
@@ -41,7 +42,7 @@ export class LocationApiService {
   public getCityName(lat: number, lng: number) {
     const url = `${this.apiUrl}/reverse?lat=${lat}&lon=${lng}&format=json&addressdetails=1&accept-language=en`;
     return this.httpClient
-      .get<NominatimResponseCity>(url)
+      .get<NominatimApiCityResponse>(url)
       .pipe(map((response) => response.address?.city || response.address?.town || 'City not found'));
   }
 }
