+++ src/environments/environment.ts
@@ -1,4 +1,3 @@
 export const environment = {
-  baseLocationApiUrl: 'http://api.openweathermap.org/geo/1.0',
-  locationApiKey: import.meta.env.NG_APP_API_LOCATION_KEY,
+  baseLocationApiUrl: 'https://nominatim.openstreetmap.org',
 } as const;
